Do not rely on a global state
Tests are failing due to state leaking between test cases.

As far as I noticed there is no need for providers to be singletons, so I suggest we pass dependencies explicitly.

This way we're not only going to prevent hard-to-track test failures, but also make the code easier to follow.

By passing dependencies explicitly, it's easy to see where various dependencies are used and how they're initialised by looking in the Bootstrap.

Signed-off-by: Jakub Zalas <jakub@zalas.pl>